♻️ Refactor: 아이콘 통일, utils 메서드 정리
Related to: #305
