+++ app/src/main/java/com/mongmong/namo/data/repositoriyImpl/ImageRepositoryImpl.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import android.net.Uri
 import com.mongmong.namo.data.datasource.s3.ImageDataSource
 import com.mongmong.namo.data.utils.RequestConverter
+import com.mongmong.namo.data.utils.RequestConverter.getFileNameFromUri
 import com.mongmong.namo.domain.repositories.ImageRepository
-import com.mongmong.namo.presentation.utils.FileUtils.getFileNameFromUri
 import okhttp3.RequestBody
 import javax.inject.Inject
