--- app/src/main/res/drawable/ic_enjoy.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="18dp"
-    android:height="18dp"
-    android:viewportWidth="18"
-    android:viewportHeight="18">
-  <path
-      android:pathData="M15.63,3.457C15.247,3.074 14.792,2.77 14.292,2.563C13.791,2.355 13.255,2.249 12.713,2.249C12.171,2.249 11.634,2.355 11.134,2.563C10.633,2.77 10.178,3.074 9.795,3.457L9,4.252L8.205,3.457C7.431,2.684 6.382,2.249 5.288,2.249C4.193,2.249 3.144,2.684 2.37,3.457C1.596,4.231 1.162,5.281 1.162,6.375C1.162,7.469 1.596,8.519 2.37,9.292L3.165,10.087L9,15.922L14.835,10.087L15.63,9.292C16.013,8.909 16.317,8.455 16.525,7.954C16.732,7.453 16.839,6.917 16.839,6.375C16.839,5.833 16.732,5.296 16.525,4.796C16.317,4.295 16.013,3.84 15.63,3.457Z"
-      android:fillColor="#DA6022"/>
-</vector>
